requirements.txt
psycopg2-binary==2.8.6
Flask==0.12.5
Flask-SQLAlchemy==2.5.1